use float instead of float64

+++ cdm_reader_mapper/cdm_mapper/properties.py
@@ -43,7 +43,7 @@ pandas_dtypes["from_atts"].update({x: x for x in numeric_types})
 # ...from CDM table definitions psuedo-sql(...) --------------------------------
 pandas_dtypes["from_sql"] = {}
 pandas_dtypes["from_sql"]["timestamp with timezone"] = "object"
-pandas_dtypes["from_sql"]["numeric"] = "float64"
+pandas_dtypes["from_sql"]["numeric"] = "float"
 pandas_dtypes["from_sql"]["int"] = "int"
 
 # Some defaults ---------------------------------------------------------------
